training/requirements.txt
torch
torchvision
tqdm
numpy
pillow
scikit-learn
matplotlib
kaggle
onnx